requirements.txt
requests
python-dotenv
pandas
numpy
coinbase
scikit-learn
streamlit
plotly
textblob
tweepy
